try thin client
