+++ src/moomoo/version
@@ -1,1 +1,1 @@
-0.3.1
+0.4.0
